linux: Move create/map window to oc_window_create
Just a check-in before deleting all the current X11 handling code. We've
reached the point where we need to interact with Mesa's EGL
implementation and this requires using XCB. Now that we've gotten more
familiar with how the X11 protocol works, I feel better using the
standard libraries for that.