It seems now we're using these checksums in two situations: 1. In CI, when we build angle, we generate checksum from the new build. The next step of the gh workflow downloads this build, verifies the checksum, and uses the build to link the runtime / package the final artifact. 2. If a contributor need the angle binary locally, but don't want to build it themselves, they can trigger a CI run, and download the artifact. The checksum get verified later when building the runtime.
It seems pointless to verify checksum of something we just built (or pulled from a gh action cache)  in 1. I'm not sure it's really useful to verify them in case 2 either, because if the artifact can be compromised, so can be the checksums (note the checksums are bundled with the build, not pulled from somewhere else), and anyway the artifacts users will get are those built in CI, not the ones we build locally while developing.

We can certainly devise a scheme later to help users verify that the artifacts they downloaded from GH are correct, but I don't think our current use of checksums does anything in this regard. The proper way would probably be to _sign_ the artifacts and verify the signature.

This commit removes said checksums until we figure out such a scheme.