remove unnamed enum
Gives `OC_INPUT_TEXT_BACKING_SIZE` a proper `typename` so it may be handled
easier by bindings generators.

This enum only exists because MSVC doesn't support using a `const` as an array size.
It can be converted into a proper `const` value when the zig build system is merged,
as `zig cc` (clang) supports this feature.